rename Assurance plugin
