+++ packages/assurance/js/index.d.ts
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 
 */
-export class AEPAssurance{
+export class Assurance{
     static extensionVersion(): Promise<string>;
     static startSession(url: string):void;
 };
